Add antialiasing: soft limiter, 2x oversampling, polyBLEP oscillators
Replace BrickWallLimiter with SoftLimiter (tanh-based) to eliminate hard
clipping discontinuities in the global effects chain. Add 2x oversampling
to TubeSaturation and Waveshaper to reduce aliasing from nonlinear
processing. Switch saw/square oscillators to polyBLEP and remove the
20-harmonic cap, adding Gibbs taper for smooth spectral rolloff on
triangle waves. Fix f32/f64 trait mismatches in example Instrument impls.
Add validation section to CLAUDE.md requiring example compilation checks.

+++ examples/hihat2.rs
@@ -51,11 +51,11 @@ const PARAM_INFO: [ParamInfo; 4] = [
 struct SharedHiHat2(Arc<Mutex<HiHat2>>);
 
 impl Instrument for SharedHiHat2 {
-    fn trigger_with_velocity(&mut self, time: f32, velocity: f32) {
+    fn trigger_with_velocity(&mut self, time: f64, velocity: f32) {
         self.0.lock().unwrap().trigger_with_velocity(time, velocity);
     }
 
-    fn tick(&mut self, current_time: f32) -> f32 {
+    fn tick(&mut self, current_time: f64) -> f32 {
         self.0.lock().unwrap().tick(current_time)
     }
